First attempt with the definition of the DAG structure
A class is created to define a DAG node, storing its type (enumeration), a reference to the information contained by that node (an index for other data structures) and the indexes for its left and right nodes. All the nodes are stored and handled with a vector, expecting to only have addition of new nodes with no care for their order in the vector.