Working over the DAG building algorithm
Added the code to handle the addition of a segment when it lays entirely inside a trapezoid (new segment crosses just one trapezoid). Surely not doing it in the smartest way, or maybe not even right (lot of doubts about this implementation). Just trying to do it, to help realize what could be ok and what not.