Fixing new errors in the add segment algorithm
The new segment was previously considered always ordered from left to right, but it wasn't: added its ordering before the start of the algorithm.
The neighbors of the new trapezoids were not updated with the new IDs of the new trapezoids, now (maybe) they are.
Fixed references to wrong named variables.