Trying to improve the trapezoid data structure
Adding some documentation and using better names for the variables and methods in the Trapezoid class. Creating a drawable version of the trapezoid (that does not extend a drawable object, because it will be drawn by the drawable trapezoidal map draw() method anyway). The drawable trapezoid stores the vertices of the trapezoid and its color. Its vertices and color will be computed by the drawable trapezoidal map (TO DO).