Adapting the trapezoidal map and the drawable trapezoidal map
Starting to adapt the trapezoidal map data structures to the previous changes.
Documented the trapezoidal map.
Declared a vector of drawable trapezoids in the drawable trapezoidal map to shadow the vector of normal trapezoids of the normal trapezoidal map. Declared some methods as virtual to correctly refer to this new vector.
Added a reference to the trapezoidal map dataset in the trapezoidal map (to decrease the number of parameters in the query and build algorithms).