Trying to improve the query and building algorithms
Added a function to initialize the DAG and the trapezoidal map, moving the responsibility of adding the initial bounding box trapezoid and leaf from the constructors of the two separate data structures to this one and only function.
Improving the readability of the query functions.
Trying to improve the readability of the building algorithm, splitting its operations in more than one function (still have to adapt the code to the previous changes).